requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
Jinja2==3.1.4

pytesseract==0.3.13
Pillow==10.4.0
pdf2image==1.17.0
opencv-python==4.10.0.84
numpy==2.1.1

scikit-learn==1.5.2
pydantic==2.8.2

requests==2.32.3
regex==2024.9.11

